Refactored helpers.js part2
