+++ EventsExpress/ClientApp/src/components/contactAdmin/contactAdmin-component.js
@@ -3,15 +3,32 @@ import { reduxForm, Field, getFormValues } from 'redux-form';
 import { connect } from 'react-redux';
 import Button from '@material-ui/core/Button';
 import { renderTextField, renderTextArea  } from '../helpers/form-helpers';
-import Module from '../helpers';
 import ErrorMessages from '../shared/errorMessage';
 import issueTypeEnum from '../../constants/IssueTypeEnum ';
 
-
-const { validate } = Module;
+const validate = values => {
+    const errors = {};
+    const requiredFields = [
+        'email',
+        'title',
+        'description'
+    ];
+    requiredFields.forEach(field => {
+        if (!values[field]) {
+            errors[field] = 'Required'
+        }
+    });
+    if (values.title && values.title.length > 30) {
+        errors.title = `Title should be less 30 symbols`;
+    }
+    if (values.email &&
+        !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
+        errors.email = 'Invalid email address'
+    }
+    return errors;
+}
 
 class ContactAdmin extends Component {
-
 
     render() {
         const { pristine, reset, submitting, error } = this.props;
@@ -50,7 +67,6 @@ class ContactAdmin extends Component {
                                 <option value={issueTypeEnum.BadUser}>Bad User</option>;
                                 <option value={issueTypeEnum.Other}>Other</option>;
                              </Field>
-
 
                             {(this.props.form_values !== undefined
                                 && this.props.form_values.subject == issueTypeEnum.Other
@@ -97,8 +113,6 @@ const mapStateToProps = (state) => {
         form_values: getFormValues('ContactAdmin')(state),
     }
 }
-
-
 
 export default connect(mapStateToProps)(reduxForm({
     form: "ContactAdmin",
