--- EventsExpress/ClientApp/src/components/helpers/validators/form-validators.js
@@ -1,2 +1,0 @@
-export const minLength = min => value =>
-    value && value.length < min ? `Must be ${min} characters or more` : undefined;
