--- EventsExpress/ClientApp/src/components/helpers/validators/index.js
@@ -1,4 +1,0 @@
-import * as FormValidators from './form-validators';
-
-export const minLength10 = FormValidators.minLength(10);
-export const minLength20 = FormValidators.minLength(20);
