+++ EventsExpress/ClientApp/src/components/inventory/visitorTakeItem.js
@@ -2,10 +2,6 @@ import React, { Component } from 'react';
 import { Field, reduxForm } from 'redux-form';
 import { renderTextField } from '../helpers/form-helpers';
 import IconButton from "@material-ui/core/IconButton";
-import Module from '../helpers';
-
-const { validate } = Module;
-
 
 class VisitorEditItemForm extends Component {
     maxValue = max => value =>
@@ -20,11 +16,11 @@ class VisitorEditItemForm extends Component {
         const { initialValues, alreadyGet } = this.props;
         return (
             <form onSubmit={this.props.handleSubmit}  className="form-inline w-100">
-                <div className="col col-md-4">
+                <div className="col col-md-3">
                     {initialValues.itemName}
                 </div>
                 <div className="col">{alreadyGet}</div>
-                <div className="col col-md-2 d-flex align-items-center">
+                <div className="col col-md-1 d-flex align-items-center">
                     <Field
                         name="willTake"
                         type="number"
@@ -33,10 +29,10 @@ class VisitorEditItemForm extends Component {
                         label="Will take"
                         component={renderTextField}/>
                 </div>
-                <div className="col col-md-1">
+                <div className="col col-md-2">
                     {initialValues.needQuantity}
                 </div>
-                <div className="col col-md-1">
+                <div className="col col-md-2">
                     {initialValues.unitOfMeasuring.shortName}
                 </div>
                 <div className="col col-md-2">
@@ -54,6 +50,5 @@ class VisitorEditItemForm extends Component {
 
 export default reduxForm({
     form: 'will-take-item-form',
-    validate: validate,
     enableReinitialize: true
 })(VisitorEditItemForm);
