+++ EventsExpress/ClientApp/src/components/login/login.js
@@ -2,7 +2,6 @@ import React, { Component } from "react";
 import { Field, reduxForm } from "redux-form";
 import DialogActions from "@material-ui/core/DialogActions";
 import Button from "@material-ui/core/Button";
-import Module from '../helpers';
 import GoogleLogin from '../../containers/GoogleLogin';
 import LoginFacebook from '../../containers/FacebookLogin';
 import TwitterLogin from '../../containers/TwitterLogin';
@@ -11,7 +10,23 @@ import { connect } from 'react-redux';
 import { Redirect } from "react-router";
 import { renderTextField } from '../helpers/form-helpers';
 
-const { validate } = Module;
+const validate = values => {
+    const errors = {}
+    const requiredFields = [
+        'email',
+        'password',
+    ]
+    requiredFields.forEach(field => {
+        if (!values[field]) {
+            errors[field] = 'Required'
+        }
+    })
+    if (values.email &&
+        !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
+        errors.email = 'Invalid email address'
+    }
+    return errors
+}
 
 class Login extends Component {
