+++ EventsExpress/ClientApp/src/components/recoverPassword/recoverPassword.js
@@ -2,12 +2,25 @@
 import { Field, reduxForm } from "redux-form";
 import DialogActions from "@material-ui/core/DialogActions";
 import Button from "@material-ui/core/Button";
-import Module from '../helpers';
 import DialogContentText from '@material-ui/core/DialogContentText';
 import ErrorMessages from '../shared/errorMessage';
 import { renderTextField } from '../helpers/form-helpers';
 
-const { validate } = Module;
+const validate = values => {
+    const errors = {}
+    const requiredFields = [
+        'email'
+    ]
+    requiredFields.forEach(field => {
+        if (!values[field]) {
+            errors[field] = 'Required'
+        }
+    })
+    if (values.email && !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
+        errors.email = 'Invalid email address'
+    }
+    return errors
+}
 
 class RecoverPassword extends React.Component {
