+++ EventsExpress/ClientApp/src/components/register/register.js
@@ -1,11 +1,36 @@
 import React, {Component} from "react";
 import DialogActions from "@material-ui/core/DialogActions";
-import Module from '../helpers';
 import {Field, reduxForm} from "redux-form";
 import Button from "@material-ui/core/Button";
-import {minLength2, maxLength15} from '../helpers/helpers'
+import { minLength6, maxLength15 } from '../helpers/validators/min-max-length-validators'
 import { renderTextField } from '../helpers/form-helpers';
-const {validate} = Module;
+
+const validate = values => {
+    const errors = {};
+    const requiredFields = [
+        'email',
+        'password',
+        'RepeatPassword',
+    ];
+    requiredFields.forEach(field => {
+        if (!values[field]) {
+            errors[field] = 'Required'
+        }
+    });
+    if (values.email &&
+        !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(values.email)) {
+        errors.email = 'Invalid email address'
+    }
+
+    if (values.password !== values.RepeatPassword) {
+        errors.RepeatPassword = 'Passwords do not match';
+    }
+
+    if (values.newPassword !== values.repeatPassword) {
+        errors.repeatPassword = 'Passwords do not match';
+    }
+    return errors;
+}
 
 class Register extends Component {
 
@@ -28,7 +53,7 @@ class Register extends Component {
                             component={renderTextField}
                             label="Password:"
                             type="password"
-                            validate={[maxLength15, minLength2]}
+                            validate={[maxLength15, minLength6]}
                         />
                     </div>
                     <div>
@@ -37,7 +62,7 @@ class Register extends Component {
                             component={renderTextField}
                             label="Repeat password:"
                             type="password"
-                            validate={[maxLength15, minLength2]}
+                            validate={[maxLength15, minLength6]}
                         />
                     </div>
                     <div>
