+++ EventsExpress/ClientApp/src/containers/edit-event-from-parent.js
@@ -8,7 +8,7 @@ import edit_event_from_parent, {
 }
     from '../actions/event/event-copy-with-edit-action';
 import * as moment from 'moment';
-import { validate } from '../components/helpers/helpers'
+import { validate } from '../components/helpers/validators/event-edit-validate-form '
 import { validateEventForm } from '../components/helpers/validators/event-validate-form'
 import get_categories from '../actions/category/category-list-action';
 import Button from "@material-ui/core/Button";
