+++ EventsExpress/ClientApp/src/containers/edit-event.js
@@ -5,7 +5,7 @@ import { connect } from 'react-redux';
 import { getFormValues, reset } from 'redux-form';
 import { setEventPending, setEventSuccess, edit_event } from '../actions/event/event-add-action';
 import { setSuccessAllert } from '../actions/alert-action';
-import { validate } from '../components/helpers/helpers'
+import { validate } from '../components/helpers/validators/event-edit-validate-form '
 import { validateEventForm } from '../components/helpers/validators/event-validate-form'
 import get_categories from '../actions/category/category-list-action';
 import Button from "@material-ui/core/Button";
